Code demographic variables in responses

Given a dataframe of survey responses, look for standard SurveyMonkey demographic questions and code/recode them into new variables that are more banner-, plot-, and analysis-friendly. Use code_std_demo_vars() to perform all standard demographic recodes, or select the specific function you need. Run std_demo_vars() to return the full index of SurveyMonkey's standard demographic variables.

Usage

code_income_vars(data, append = TRUE, verbose = TRUE)

Arguments

data

Input data frame of survey responses.

append

If TRUE, append to existing dataframe otherwise return new variables in a fresh dataframe.

verbose

Print helpful messages.

Value

Survey data frame with new (or updated) columns corresponding to (if discoverable) respondents' income levels:

income: original income scale

income3: 3-point income scale (i.e. <$50k, $50-100k, >$100k)