Browse user's surveys

Retrieve a dataframe of the user's surveys.

Usage

browse_surveys(
  per_page = 100,
  page = NULL,
  sort_by = NULL,
  sort_order = NULL,
  start_modified_at = NULL,
  end_modified_at = NULL,
  title = NULL,
  include = NULL,
  folder_id = NULL,
  oauth_token = get_token()
)

Arguments

per_page

Integer number to set the number of surveys to return per page. Maximum value is 1000 surveys per page; try that if your survey is not on the first 100, to reduce API calls.

page

Integer number to select which page of resources to return. By default is 1.

sort_by

String used to sort returned survey list: ‘title’, 'date_modified’, or 'num_responses’. By default, date_modified.

sort_order

String used to set the sort order for returned surveys: 'ASC’ or 'DESC’. By default, DESC.

start_modified_at

Date string used to select surveys last modified after this date. By default is NULL.

end_modified_at

Date string used to select surveys modified before this date. By default is NULL.

title

String used to select survey by survey title. By default is NULL.

include

Character vector as a comma separated string used to filter survey list: 'response_count’, 'date_created’, 'date_modified’, 'language’, 'question_count’, 'analyze_url’, 'preview’. By default is NULL. Use browse_surveys('everything') to pull all fields.

folder_id

Specify the id of a folder to only return surveys in it.

oauth_token

Your OAuth 2.0 token. By default, retrieved from get_token().

Value

A dataframe of surveys and associated metadata from desired account.

Details

This function calls the SurveyMonkey API using the current OAuth token and returns a dataframe of surveys filtered by the parameters entered.

References

SurveyMonkey API V3 at https://developer.surveymonkey.com/api/v3/#surveys.